Separete config view class
Separate SDKConfigView class to single file.
And also some functions to be a common.
